Added autosave to title bar (with opt. fix)
Added TimeMonitor class for measuring lengths of time; used for autosave functionality without slowing the program down to a crawl. 
Now, you must stop typing in the title field for at least 0.55 seconds for the title to be actually saved to the file. 
Added updates to the project settings page; starting work on the starred/tags/pictures/attachedFiles/linkedItems